+++ package.json
@@ -11,8 +11,9 @@
     "tsToJs": "rm -rf ./src && tsc --skipLibCheck",
     "tsToJsOrTrue": "rm -rf ./src && tsc --skipLibCheck || true",
     "tsEslint": "node_modules/eslint/bin/eslint.js 'src_ts/**/*.ts' --fix",
-    "start": "tsc || true && concurrently -k -r \"tsc --watch --preserveWatchOutput\" \"wds --hmr\"",
-    "build": "export NODE_OPTIONS=--max_old_space_size=4096 && rollup -c rollup.config.js",
+    "start": "rm -rf src && rollup -c rollup-dev.config.js -w",
+    "start:tsc": "tsc || true && concurrently -k -r \"tsc --watch --preserveWatchOutput\" \"wds --hmr\"",
+    "build": "export NODE_OPTIONS=--max_old_space_size=4096 && rollup -c rollup-prod.config.js",
     "test": "npm run test:unit && npm run test:integration",
     "test:integration": "tsc && mocha test/integration --timeout=10000",
     "test:regenerate_screenshots": "tsc && mocha test/integration/screenshots-baseline/regenerate.js --timeout=15000",
@@ -63,7 +64,7 @@
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
     "@typescript-eslint/typescript-estree": "^5.7.0",
-    "@web/dev-server": "^0.1.35",
+    "esbuild": "^0.19.5",
     "ajv": ">=6.12.6",
     "async": ">=2.6.4",
     "chokidar-cli": "^3.0.0",
@@ -91,9 +92,12 @@
     "relative-deps": "^1.0.7",
     "rollup": "^2.79.1",
     "rollup-plugin-copy": "^3.4.0",
+    "rollup-plugin-esbuild": "^6.1.0",
     "rollup-plugin-delete": "^2.0.0",
     "rollup-plugin-filesize": "^9.1.2",
     "rollup-plugin-license": "^3.2.0",
+    "rollup-plugin-livereload": "^2.0.5",
+    "rollup-plugin-serve": "^2.0.2",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-workbox": "^8.0.1",
     "shelljs": ">=0.8.5",
