+++ rollup-dev.config.js
@@ -1,0 +1,56 @@
+import defaultConfig from './rollup.config.js';
+import serve from 'rollup-plugin-serve';
+import livereload from 'rollup-plugin-livereload';
+import copy from 'rollup-plugin-copy';
+
+// For livereload to work you need to expose port in docker-compose.yaml with ports: -"3001:3001"
+const appPort = 8080;
+const liveReloadPort = 3002;
+
+console.log("\x1b[92m" + "Starting app on port:" + appPort + " \x1b[0m");
+console.log("\x1b[92m" + "LiveReload starting on port:" + liveReloadPort + " \x1b[0m");
+console.log("\x1b[93m" + "LiveReload Hint: If live reload is not working you probably must expose port " + liveReloadPort + " in docker-compose.yaml" + " \x1b[0m");
+
+// Extra files to copy in src directory ./src
+const copyConfig = {
+  targets: [
+    {
+      src: 'src_ts/**/*.json',
+      dest: 'src'
+    }
+  ],
+  copyOnce: true,
+  flatten: false
+};
+
+
+const config = {
+  ...defaultConfig,
+  output: {
+    dir: 'src',
+    format: 'es',
+    sourcemap: true,
+    entryFileNames: '[name].js',
+    chunkFileNames: '[name].js',
+  },
+  watch: {
+    include: 'src_ts/**'
+  },
+  plugins: [
+    ...defaultConfig.plugins,
+    copy(copyConfig),
+    serve({
+      historyApiFallback: true,
+      openPage: 'index.html',
+      port: appPort
+    }),
+    livereload({
+      port: liveReloadPort,
+      verbose: true,
+      delay: 1000,
+      watch: 'src',
+    }),
+  ]
+}
+
+export default config;
